+++ common/src/main/java/io/github/flemmli97/simplequests/player/PlayerData.java
@@ -153,6 +153,11 @@ public class PlayerData {
                 (prog, p) -> this.player.sendMessage(new TranslatableComponent(ConfigHandler.lang.get("simplequests.task"), p.getSecond().translation(this.player.getServer())).withStyle(ChatFormatting.DARK_GREEN), Util.NIL_UUID));
     }
 
+    public void onItemCrafted(ItemStack stack, int amount, String trigger) {
+        this.tryFullFill(QuestEntryImpls.CraftingEntry.class, QuestProgress.createCraftingPredicate(this.player, stack, amount),
+                (prog, p) -> this.player.sendMessage(new TranslatableComponent(ConfigHandler.lang.get("simplequests.task"), p.getSecond().translation(this.player.getServer())).withStyle(ChatFormatting.DARK_GREEN), Util.NIL_UUID), trigger);
+    }
+
     private void completeQuest(QuestProgress prog) {
         LootTable lootTable = this.player.getServer().getLootTables().get(prog.getQuest().loot);
         CriteriaTriggers.GENERATE_LOOT.trigger(this.player, prog.getQuest().loot);